Update LostFoundAndStolen
Added scoring logic, but broke unhideSecret [Bandits stay hidden when checked] and checkForWinner() [Treasure doesnt point to winning face]:
each Bandit has a one time use secret number of 2,4, and 6
the secret number used by a  Bandit in a round, win or lose, cannot be used again
ex: on the first round I use a 4 and someone else plays a 6, I cannot play 4 using this Bandit next round.  If they play a 2 my Bandit becomes a TeamTreasure score keeper.  If we both play 4 then we both take back the losing Bandit and neither can use a 4 on that Bandit again.  However, depending on the number of Bandit blinks in our hands, we could still have another Bandit that could play a 4.