Update LostFoundAndStolen
Face 0 is only face that Bandit sends secret number on, so we can choose which Treasure to go for.  We can now steal 3 from Treasure always, steal 1 from TeamTreasure with 3 or 2 left, and if we try to steal from a TeamTreasure with only 1 remaining we waste our Bandit's secret number.