Update LostFoundAndStolen
New scoring working.
Only highest unique secret number wins. Other Bandits played lose that number for that Blink.
If Bandit wins with a 2, then their awarded treasure is locked/ blocked from stealing.  If a Bandit wins with a 4 or 6 the treasure can be stolen.  
Stealing: if 3 treasure, then steal 2. If 2 treasure then steal 1, if 1 treasure then steal the last 1 and that Bandit is now empty.
Only face 0, designated by white blinking on Bandit when not in hide secret mode, sends secret signal.  Placing other faces on treasure can be used as bluffing.  
Gameplay:
1. divide up Blinks evenly keeping 1 separate for the initial Treasure.
2. Players take turns placing Bandits on the board until they dont want to (could alway bluff) or all Treasure faces are occupied.
3. Winning Bandit blinks Find or Steal treasure
4. play until all Bandits are used up or a Band of Bandits (team) scores 10 points